reddit_scraper/requirements.txt
asyncpraw
httpx
python-dotenv